Lots of misc drilling improvements; zoom and bubble charts based on d3 and highstocks, respectively.
This is a little warty, I didn't have the time I wanted to clean things up.  But it works, and with a little demonstration, it works fairly well.  There are more directions I wanted to take this - for instance, it would be handy to have the ability to compare detail drill downs side by side.